Revert package rename
